requirements.txt
Django==4.2
#static file uchun
whitenoise==6.6.0
#hosting uchun 
dj-database-url==2.1.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
django-environ==0.11.2
pillow==10.4.0
#back end uchun
djangorestframework==3.15.1
requests==2.31.0
django-cors-headers==4.3.1
openpyxl==3.1.2
python-environ==0.4.54
flower==2.0.1
celery
django-redis
